Added protanomaly/deuteranomaly/tritanomaly simulation to the shader.
The simulation is likely inaccurate, but I figure it doesn't hurt
to have it in the shader code (which means it can be forced on with
a manual config file change).

I of course am not an anomalous trichromat, but looking from a
theoretical point of view, the results I see are way off from what
I'd expect. I find it surprising that one of the things it does is
severely oversaturate the color blue, for instance, which makes
desaturated blues look shockingly bright (most notably, the hyperspace
connection lines on the starmap).