Invalid colorblindness type: also include blue component.
Rods aren't sensitive to red wavelengths, so it's possible that someone
with rod monochromacy, in the error condition for colorblind simulation,
might be able to see little or none of the display with an all-red
error condition, which would hinder their ability to correct it let
alone play the game. Adding a blue component (and thus making the
error condition purple) should correct this (and since it's an error
condition, the exact hue doesn't otherwise matter).