Scrapped non-linear energy regeneration, changing to simple linear.
Funny enough, we tried to do this to Naev *years* ago, but got
pushback. And now bobbens is immediately giving up on it the second
someone else makes a PR for it. Go figure.

Well, it looks like we're in agreement on the issue now: the whole
"realistic" non-linear energy regeneration thing was a failure. It
also *wasn't* realistic, because energy used by weapons and activated
outfits has always been linear, entirely bypassing the formula used
for energy regeneration, and that was because that formula for energy
regeneration didn't work properly in reverse. So the actual effect
was to make energy regeneration lower than advertised.

Not doing the whole hack of still making energy regeneration lower
than advertised that was done on Naev, so this may mean a need for
changes to some stats.