Redid the weapon heat tutorial, incorporating changes to the Llama.
I've made it so that the Llama has a huge fire rate boost, but a
hilariously bad heat increase, for forward weapons, but with a small
bonus to heat dissipation. I've also increased the health of the
Hakoi asteroids. These changes conspire to essentially guarantee that
the player's weapons will overheat while mining for tutorial3, and
so the after-mining cooldown step has been replaced with a gentle
suggestion from Ian Structure to maybe cool the ship down when the
weapons start to get badly overheated.