Fixed map trade mode descriptions.
There were a couple of problems. Firstly, the colors instructed were
wrong. Secondly, a newline caused one of them to bleed into the buttons
because that overrode the automatic breaking. Third, that one was
also way too long. I also changed the color descriptions with just
colored hilights, which I think is better.