Don't allow static event pilots to spawn too close to planets.
Affects the derelict, distress, and shipwreck events. I also changed
the way the shipwreck event chooses the location from 2000–3000 mAU
from the center to between the center and half the system's radius.
I'm guessing the original positioning was based on way back when Naev
followed EV's approach of having just one planet in a fairly small
system; it was probably assuming that the center of the system was
"the planet", which hasn't ever been the case in Naikari and hasn't
been the case in Naev for over a decade now.