Got started on more damage type conversions.
The shield→armor and armor→shield conversions aren't implemented yet
because it turns out this is going to be way harder than I would've
thought, due to the reliance on static damage types to manipulate this.
Also, I'm not entirely sure about the way to display these stats to
the user.