Fully implemented damage conversions, with associated necessary code tweaks.
Had to change Damage structs to contain actual damage modifier info
in addition to damage type, which is now used by dtype_calcDamage for
the actual damage calculation, so that weapon damage can be modified
at the shield, armor, and knockback components.

I also noticed that none of the stats (including damage as disable)
were affecting asteroid weapon collisions and fixed that. Might have
been overlooked by bobbens because when you only have "damage as disable"
failing to do the conversion becomes a benefit with asteroid mining,
but I want ionized weapons to become weaker when firing on asteroids
and deionized weapons to become stronger.