Decided how to implement bolt weapon charging.
Instead of just a flat "charged" flag, I've changed it to a
"charge capacity" property (which we'll allow to be modified by stats
later) indicating how many seconds of charge the weapon can take.
Charging will happen in the background if the weapon is not firing
*and* not in cooldown, and its damage will be increased based on the
amount of charge the weapon has, then charge will be reset to zero
when the weapon is fired.