Added support for range definitions for certain stats.
These are display ranges used only for the "composite stats display on a
ship, with their purpose being to visually show to the player when
a stat has reached its effective cap. This is going to become important
as we start allowing much more freeform outfit equipping; you could,
for example, have several outfits that add up to +392° shot spread,
which is functionally identical to +360° and thus should be displayed
as such.