Completely removed planet and bar descriptions.
To explain why, these descriptions, while they can help a little with
worldbuilding, mostly end up being a bunch of busywork that adds to
development time with little difference to the player. The idea of
coming up with hundreds of interesting planet descriptions, which
also have to be tailored to correspond to gameplay, is also a bit
absurd; there's only so many ideas you can come up with for interesting
planet descriptions before you start falling back on telling pointless
side-stories or making shitty jokes or references. Overall, these
descriptions, I've learned from experience, inevitably turn into the
least polished aspect of the game, with no real benefit.

Instead of using paragraphs to tell the player what the planets are like,
I'd prefer to use organic storytelling methods. This not only is a
nicer experience for the player, but it also makes designing these
worlds much simpler, given that arbitrary characteristics can come
together and the player can fill in the blanks with their own
imagination.