Renamed all "Empire" ships to "Imperial", changed Empire spawn code.
I wanted to make it so that 1 presence roughly equals 1 ship, rather
than presence being an entirely arbitrary number. To start this off,
I've done the adjustment for Empire spawning. While I did this, I
renamed all "Empire" ships to "Imperial" and deleted the Empire
fighter bays as well as all references to them. Also added the start
of the mechanism we'll be using going forward for drones (which will
become the only fighter bays).