Less creds, trimmed down the randomly generated names in the start event.
The starting credits have been lowered from 30,000 to just 5,000.

I've cut the list of randomly generated names down. The player will
only ever see one, so I want to make sure all of them are great, and
in particular, that they're all references that are meaningful in
some way to me personally.