Fixed a segfault caused by map commodity modes.
I find this fix to be… patchy, but it's the best way I could come
up with that doesn't involve completely tearing apart everything and
starting from scratch.

Basically, there was code that depended on an assumption that the
player knows about all commodities. That has happened to be true for
years now, but an outdated save file could cause there to be no
known planets, leading to a mismatch. I "fixed" this by making it
so that the map code always considers all "standard" commodities to
be "known".