Removed support for compressing saved games.
I didn't check this, but in looking out of of curiosity at what Naev
is up to, I noticed that the most recent update removes support
because supposedly support for compression will be dropping in
libxml at some point. Sounds good to me; save game compression was
never really a great idea to begin with.