Heavily boosted and normalized disable potential of lethal weapons.
This system for how to disable ships creates a particular quirk in
Naev, where you have a dichotomy between being able to destroy *or*
disable ships. This change is meant to reduce that dichotomy, without
completely scrapping the system.

All lethal weapons (haven't touched launchers yet) now inflict a
significant amount of disable as well, which should make it possible
for any weapon to disable a ship, as long as the player exercises
some care. Different weapons have different levels of effectiveness
at this, with Razors being the most effective and Crystals being the
least effective.