Added outfitting to tutorial3, deleted the old outfits tutorial.
Outfits and ship customization are a very central part of this game,
so it really should be introduced early on. Tutorial 3 is a perfect
fit for this, since it involves getting cargo. Now, Ian asks for a
specific amount of cargo, rather than for you to fill your capacity,
and the amount is (with some exceptions) deliberately chosen to force
you to buy a Cargo Pod, which is now available on Kikero and just
cheap enough to be able to buy it after completing Ian's first
mission.