Consolidated shield and energy boosting outfits.
This removes the Shield Booster and Reactor outfits, and instead
adds those effects to the Shield Capacotor and Battery outfits.
Solar Panel is also once again a structural, and enhanced slightly
to make it more useful.

I also renamed several outfits to make things more consistent and
easy to change with a find-and-replace.

I'm thinking of renaming the Battery outfits to something else, but
don't have a better name in mind yet.