Removed the "fighter bay" slot property.
This was added at a time when I thought it was important to
differentiate different ships a bit more. Now, I believe that there
actually should be almost no difference at all between ships of the
same size. The vision of this project is to turn each ship into a
canvas that they can customize, so restrictions like this are a
hindrance to the goal of the game's design.

This is doubly true because fighter bays will become exclusive to
certain factions.

The ship descriptions will still boast about unique features, but
I'm thinking the way I'll create the appearance of differentiation
is actually just different initial outfits, plus some very minor
stat differences. Haven't gone into this yet.