Pyxidis → Virginis
This was the name I originally wanted to give the starting systems,
but I chose the less good "Pyxidis" out of a worry that people would
make inappropriate jokes about the name "Virginis". But "Virginis"
is a good fit for this constellation because it's where the player
is new and learning how to play.